Refactoring to split out reusable bits from web-server into lib-web
