+++ crates/libs/lib-web/src/handlers/mod.rs
@@ -1,0 +1,2 @@
+pub mod handlers_login;
+pub mod handlers_rpc;
