+++ crates/libs/lib-web/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod error;
+pub mod log;
+pub mod handlers;
+pub mod middleware;
+pub mod routes;
+pub mod utils;
