+++ crates/libs/lib-web/src/routes/routes_static.rs
@@ -1,4 +1,3 @@
-use crate::web_config;
 use axum::handler::HandlerWithoutStateExt;
 use axum::http::StatusCode;
 use axum::routing::{any_service, MethodRouter};
@@ -6,13 +5,13 @@ use tower_http::services::ServeDir;
 
 // Note: Here we can just return a MethodRouter rather than a full Router
 //       since ServeDir is a service.
-pub fn serve_dir() -> MethodRouter {
+pub fn serve_dir(web_folder: &'static String) -> MethodRouter {
 	async fn handle_404() -> (StatusCode, &'static str) {
 		(StatusCode::NOT_FOUND, "Resource not found.")
 	}
 
 	any_service(
-		ServeDir::new(&web_config().WEB_FOLDER)
+		ServeDir::new(web_folder)
 			.not_found_service(handle_404.into_service()),
 	)
 }
