+++ crates/libs/lib-web/src/utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod token;
